pyproject.toml
[project]
name = "FastAuth"
version = "0.1.0"
description = "A comprehensive FastAPI package for role-based access control with support for multiple auth providers"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.104.0",
    "pydantic[email]>=2.0.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "python-multipart>=0.0.6",
    "httpx>=0.25.0",
    "sqlalchemy>=2.0.0",
    "alembic>=1.12.0",
]

[project.optional-dependencies]
auth0 = ["auth0-python>=4.9.0"]
firebase = ["firebase-admin>=6.0.0"]
redis = ["redis>=5.0.0"]
sqlite = []  # SQLite is included with Python
postgres = ["psycopg2-binary>=2.9.0"]
mysql = ["pymysql>=1.1.0"]
oracle = ["cx-oracle>=8.3.0"]
mongodb = ["motor>=3.3.0", "pymongo>=4.6.0"]
all-databases = [
    "psycopg2-binary>=2.9.0",
    "pymysql>=1.1.0", 
    "cx-oracle>=8.3.0",
    "motor>=3.3.0",
    "pymongo>=4.6.0"
]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
    "mypy>=1.5.0",
    "pre-commit>=3.4.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["fastauth"]

[tool.black]
line-length = 88
target-version = ['py312']

[tool.isort]
profile = "black"
multi_line_output = 3

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
disallow_any_unimported = true
disallow_any_expr = false
disallow_any_decorated = false
disallow_any_explicit = false
disallow_any_generics = true
disallow_subclassing_any = true
disallow_untyped_calls = true
disallow_untyped_decorators = true
disallow_incomplete_defs = true
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true
strict_concatenate = true
show_error_codes = true
show_column_numbers = true
pretty = true

[[tool.mypy.overrides]]
module = [
    "jose.*",
    "passlib.*",
    "auth0.*",
    "firebase_admin.*",
    "redis.*",
    "psycopg2.*",
    "pymysql.*",
    "alembic.*",
]
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "--cov=fastauth --cov-report=term-missing"

[dependency-groups]
dev = [
    "uvicorn>=0.34.2",
]
